requirements.txt
Django==4.0
Pillow==9.5.0
django-admin-sortable2==2.1.8